Add host and port command line arguments
The service now supports taking the host and port as command line
arguments.

The Dockerfile has been changed to pass the 0.0.0.0 host in the CMD
instruction while the ENTRYPOINT points to the service script. In
addition the large dependencies have been inlined in the Dockerfile to
take advantage of image caching.

Signed-off-by: Byron Ruth <b@devel.io>